Fuel consumption feature

+++ app/src/main/java/com/freed_asd/fuel_calculator/data/local/consumption/mixed/ConsMixed.kt
@@ -10,5 +10,5 @@ import androidx.room.PrimaryKey
 data class ConsMixed (
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "consumption")val Consumption: Float,
-    @ColumnInfo(name = "mileage")val mileage: Int
+    @ColumnInfo(name = "mileage")val mileage: Float
 )
